today space badge for all spaces
